Finished processor implemntation (except RMVPC), fixed data implementations

+++ src/main/java/municipaldata/common/PropertyValue.java
@@ -2,11 +2,11 @@ package municipaldata.common;
 
 public class PropertyValue {
 
-    float market_value;
-    float total_livable_area;
+    double market_value;
+    double total_livable_area;
     String zip_code;
 
-    public PropertyValue(float market_value, float total_livable_area, String zip_code) {
+    public PropertyValue(double market_value, double total_livable_area, String zip_code) {
         this.market_value = market_value;
         this.total_livable_area = total_livable_area;
         this.zip_code = zip_code;
@@ -16,11 +16,11 @@ public class PropertyValue {
         return zip_code;
     }
 
-    public float getMarketValue() {
+    public double getMarketValue() {
         return market_value;
     }
 
-    public float getTotalLivableArea() {
+    public double getTotalLivableArea() {
         return total_livable_area;
     }
 }
